+++ Sources/Sentry/SentryBreadcrumb.m
@@ -1,4 +1,5 @@
 #import "SentryBreadcrumb.h"
+#import "SentryBreadcrumb+Private.h"
 #import "SentryDateUtils.h"
 #import "SentryLevelMapper.h"
 #import "SentryNSDictionarySanitize.h"
@@ -30,6 +31,8 @@
                 self.category = value;
             } else if ([key isEqualToString:@"type"] && isString) {
                 self.type = value;
+            } else if ([key isEqualToString:@"origin"] && isString) {
+                self.origin = value;
             } else if ([key isEqualToString:@"message"] && isString) {
                 self.message = value;
             } else if ([key isEqualToString:@"data"] && isDictionary) {
@@ -69,6 +72,7 @@
     [serializedData setValue:sentry_toIso8601String(self.timestamp) forKey:@"timestamp"];
     [serializedData setValue:self.category forKey:@"category"];
     [serializedData setValue:self.type forKey:@"type"];
+    [serializedData setValue:self.origin forKey:@"origin"];
     [serializedData setValue:self.message forKey:@"message"];
     [serializedData setValue:sentry_sanitize(self.data) forKey:@"data"];
     NSDictionary<NSString *, id> *unknown = self.unknown;
@@ -106,6 +110,8 @@
         return NO;
     if (self.type != breadcrumb.type && ![self.type isEqualToString:breadcrumb.type])
         return NO;
+    if (self.origin != breadcrumb.origin && ![self.origin isEqualToString:breadcrumb.origin])
+        return NO;
     if (self.message != breadcrumb.message && ![self.message isEqualToString:breadcrumb.message])
         return NO;
     if (self.data != breadcrumb.data && ![self.data isEqualToDictionary:breadcrumb.data])
@@ -123,6 +129,7 @@
     hash = hash * 23 + [self.category hash];
     hash = hash * 23 + [self.timestamp hash];
     hash = hash * 23 + [self.type hash];
+    hash = hash * 23 + [self.origin hash];
     hash = hash * 23 + [self.message hash];
     hash = hash * 23 + [self.data hash];
     hash = hash * 23 + [self.unknown hash];