+++ Sources/Sentry/SentryCrashReportConverter.m
@@ -1,4 +1,5 @@
 #import "SentryCrashReportConverter.h"
+#import "SentryBreadcrumb+Private.h"
 #import "SentryBreadcrumb.h"
 #import "SentryCrashStackCursor.h"
 #import "SentryDateUtils.h"
@@ -170,6 +171,7 @@
                      category:storedCrumb[@"category"]];
             crumb.message = storedCrumb[@"message"];
             crumb.type = storedCrumb[@"type"];
+            crumb.origin = storedCrumb[@"origin"];
             crumb.timestamp = sentry_fromIso8601String(storedCrumb[@"timestamp"]);
             crumb.data = storedCrumb[@"data"];
             [breadcrumbs addObject:crumb];